[C-3211] Remove legacy upload
