+++ packages/common/src/services/remote-config/feature-flags.ts
@@ -45,7 +45,6 @@ export enum FeatureFlags {
   PROXY_WORMHOLE = 'proxy_wormhole',
   AI_ATTRIBUTION = 'ai_attribution',
   DEVELOPER_APPS_PAGE = 'developer_apps_page',
-  UPLOAD_REDESIGN_ENABLED = 'upload_redesign_enabled',
   USDC_PURCHASES = 'usdc_purchases',
   USDC_PURCHASES_UPLOAD = 'usdc_purchases_upload',
   AUDIO_MATCHING_CHALLENGES = 'audio_matching_challenges',
@@ -113,7 +112,6 @@ export const flagDefaults: FlagDefaults = {
   [FeatureFlags.PROXY_WORMHOLE]: false,
   [FeatureFlags.AI_ATTRIBUTION]: false,
   [FeatureFlags.DEVELOPER_APPS_PAGE]: false,
-  [FeatureFlags.UPLOAD_REDESIGN_ENABLED]: false,
   [FeatureFlags.USDC_PURCHASES]: false,
   [FeatureFlags.USDC_PURCHASES_UPLOAD]: false,
   [FeatureFlags.AUDIO_MATCHING_CHALLENGES]: false,
