--- packages/web/src/pages/upload-page/UploadPageNew.module.css
@@ -1,9 +1,0 @@
-.upload {
-  text-align: left;
-}
-
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
