+++ packages/web/src/pages/upload-page/components/FinishPage.module.css
@@ -1,97 +1,3 @@
-.buttons button {
-  margin-left: 16px;
-}
-
-.continueButton {
-  all: unset;
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  color: var(--static-neutral);
-  font-size: var(--font-s);
-  font-weight: var(--font-bold);
-  text-align: right;
-  transition: all ease-in-out 0.07s;
-  opacity: 1;
-  transform: scale3d(1, 1, 1);
-  backface-visibility: hidden;
-  cursor: pointer;
-}
-
-.continueButton:hover {
-  text-decoration: underline;
-  transform: scale3d(1.02, 1.02, 1.02);
-}
-
-.continueButton:active {
-  transform: scale3d(0.99, 0.99, 0.99);
-}
-
-.continueButton.isHidden {
-  opacity: 0;
-}
-
-.iconArrow {
-  margin-left: 4px;
-  height: 12px;
-  width: 12px;
-}
-
-.iconArrow path {
-  fill: var(--static-neutral);
-}
-
-.uploadingInfo {
-  width: 100%;
-  display: inline-flex;
-  align-items: center;
-}
-
-.progressBar {
-  flex: 1 0 auto;
-}
-
-.progress {
-  height: 12px;
-}
-
-.progressSlider {
-  transition: 0.2s ease-in-out;
-}
-
-.uploadText {
-  margin: 0px 16px 0px 24px;
-  flex: 0 0 70px;
-  color: var(--neutral-light-4);
-  font-size: var(--font-xs);
-  font-weight: var(--font-demi-bold);
-  text-align: center;
-  white-space: nowrap;
-  width: 80px;
-  transition: color 0.2s ease-in-out;
-}
-
-.uploadText.uploadComplete {
-  color: var(--primary);
-}
-
-.trackListContainer {
-  background: var(--white);
-}
-
-.userName {
-  display: flex;
-  align-items: center;
-}
-
-.createdBy {
-  margin-right: 4px;
-}
-
-.iconVerified {
-  margin-left: 4px;
-}
-
 .page {
   display: flex;
   flex-direction: column;
