+++ packages/web/src/pages/upload-page/components/SelectPage.module.css
@@ -7,41 +7,6 @@
   height: 600px;
   position: relative;
 }
-.uploaded {
-  flex: 1 1 0;
-  transition: flex-grow ease-in-out 0.18s;
-  margin-left: 16px;
-  max-width: 50%;
-}
-.uploaded.hide {
-  flex-grow: 0.0001;
-  margin: 0;
-}
-
-.dropdown {
-  width: 200px;
-  margin-bottom: 16px;
-}
-
-.count {
-  margin-top: 16px;
-  padding-left: 8px;
-  text-align: left;
-  user-select: none;
-
-  font-size: var(--font-xs);
-  font-weight: var(--font-medium);
-  line-height: 14px;
-}
-
-.continue {
-  text-align: right;
-}
-
-.continueButton .continueButtonText {
-  font-weight: var(--font-bold);
-  font-size: var(--font-l);
-}
 
 .invalidFileType {
   position: absolute;
@@ -51,6 +16,14 @@
   margin: auto;
 }
 
-.addTrackNotifications {
-  margin-bottom: 8px;
+.uploaded {
+  flex: 1 1 0;
+  transition: flex-grow ease-in-out 0.18s;
+  margin-left: 16px;
+  max-width: 50%;
+}
+
+.hide {
+  flex-grow: 0.0001;
+  margin: 0;
 }
