+++ packages/web/src/pages/upload-page/components/SelectPageNew.tsx
@@ -15,7 +15,7 @@ import { processFiles } from '../store/utils/processFiles'
 import { UploadFormState } from '../types'
 
 import styles from './SelectPage.module.css'
-import { TracksPreviewNew } from './TracksPreviewNew'
+import { TracksPreview } from './TracksPreviewNew'
 
 type ErrorType = { reason: 'corrupted' | 'size' | 'type' } | null
 
@@ -118,7 +118,7 @@ export const SelectPageNew = (props: SelectPageProps) => {
           })}
         >
           {tracks.length > 0 ? (
-            <TracksPreviewNew
+            <TracksPreview
               tracks={tracks}
               uploadType={uploadType}
               onRemove={onRemoveTrack}
