+++ packages/web/src/pages/upload-page/components/ShareBannerNew.tsx
@@ -55,7 +55,7 @@ export const ShareBannerNew = (props: ShareBannerProps) => {
 
   return (
     <div
-      className={styles.containerNew}
+      className={styles.root}
       style={{
         backgroundImage: `linear-gradient(315deg, rgba(91, 35, 225, 0.8) 0%, rgba(162, 47, 237, 0.8) 100%), url(${backgroundPlaceholder})`
       }}
@@ -66,7 +66,7 @@ export const ShareBannerNew = (props: ShareBannerProps) => {
       <Text variant='heading' size='medium' color='darkmodeStaticWhite'>
         {messages.shareText}
       </Text>
-      <div className={styles.buttonContainerNew}>
+      <div className={styles.buttonContainer}>
         <Button
           fullWidth
           leftIcon={<IconTwitterBird />}
