+++ packages/web/src/pages/upload-page/components/TracksPreview.module.css
@@ -1,69 +1,6 @@
-.container {
-  max-height: 600px;
-  display: inline-flex;
-  flex-direction: column;
-  width: 100%;
-  user-select: none;
-}
-
-.headerContainer {
-  width: 100%;
-  border-radius: 8px 8px 0 0;
-  background-color: var(--neutral-light-10);
-  box-shadow: 0 2px 5px -2px rgba(133, 129, 153, 0.25);
-  padding: 16px 16px 14px;
-  display: inline-flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  border-bottom: 1px solid var(--neutral-light-8);
-}
-
-.header {
-  text-transform: uppercase;
-  color: var(--neutral-light-2);
-  font-size: 15px;
-  font-weight: bold;
-  letter-spacing: 0.5px;
-  margin-bottom: 12px;
-  user-select: none;
-}
-
-.typeSelector {
-  margin-bottom: 15px;
-}
-
-.typeDescription {
-  color: var(--neutral);
-  font-family: var(--font-family);
-  font-size: var(--font-s);
-  font-weight: var(--font-medium);
-  line-height: 23px;
-  user-select: none;
-}
-
-.tracks {
-  max-height: 467px;
-  box-shadow: 0 2px 5px -2px rgba(133, 129, 153, 0.25);
-  padding: 8px 0px;
-  border-radius: 0 0 8px 8px;
-  background-color: var(--white);
-  -webkit-overflow-scrolling: touch;
-  -webkit-transform: translate3d(0, 0, 0);
-}
-
-.tracks.shortScroll {
-  max-height: 444px;
-}
-
-.tabSlider {
-  margin-bottom: 15px;
-}
-
-.containerNew {
+.root {
   display: flex;
   flex-direction: column;
-  /* TODO: KJ - Figure out if we need this  */
   max-height: 600px;
   width: 100%;
   user-select: none;
@@ -72,7 +9,7 @@
     0px 0px 4px 0px rgba(0, 0, 0, 0.04);
 }
 
-.infoContainerNew {
+.info {
   display: flex;
   flex-direction: column;
   gap: 16px;
@@ -81,21 +18,24 @@
   background-color: var(--background-surface-1);
 }
 
-.headerContainerNew {
+.header {
   border-bottom: 1px solid var(--border-strong);
   border-radius: 8px 8px 0 0;
 }
 
-.footerContainerNew {
+.footer {
   border-top: 1px solid var(--border-strong);
   border-radius: 0 0 8px 8px;
 }
 
-.tracksContainerNew {
+/* keep */
+.shortScroll {
+  max-height: 444px;
+}
+
+.tracks {
   display: flex;
   padding: 16px 0;
   flex-direction: column;
   background-color: var(--background-white);
-  /* TODO: KJ - Figure out if we need this  */
-  /* max-height: 467px; */
 }
