+++ packages/web/src/pages/upload-page/index.ts
@@ -1,0 +1,1 @@
+export { UploadPage as default } from './UploadPage'
